feat: enable relationship with parents in rulebook process
